% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLUSTShiny.R
\name{CLUSTShiny}
\alias{CLUSTShiny}
\title{Launch 'CLUSTShiny' Interface}
\usage{
CLUSTShiny()
}
\value{
Nothing
}
\description{
CLUSTShiny() loads interactive user interface built using R 'shiny'.
}
\details{
Start CLUSTShiny

The interactive user interface is to provide an easy way for cluster analysis and downloading relevant plots.
}
\examples{
if(interactive()){
library(rmarkdown)
CLUSTShiny()
}
}
\keyword{CLUSTShiny}
